instances-appspec/api-server/app-api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
slowapi
